refactor(automation_v2): extract priority logic to priority.go, soften dense-ordering claim
- Move the five priority helpers (~105 lines) out of the 1.3k-line
  resource.go into priority.go, mirroring the existing priority_test.go
  split; resource.go now grows only ~15 lines over master for this
  feature.
- Document why applyPlanPriorityOnUpdate's error path deliberately
  skips State.Set (framework keeps prior state on errored Update),
  unlike Create's explicit null-and-set.
- Schema description now says ordering is intended to be dense 1..N
  but legacy data may contain gaps or duplicates (verified live: the
  US lab has two automations at priority 2); template and generated
  docs updated to match, and the Managing Priority example comment now
  points at the required attributes.
- Put Description before Optional in the priority attribute to match
  the dominant sibling field order.
